requirements.txt
boto3==1.16.2
botocore==1.19.2
Flask==1.1.2
gunicorn==20.0.4
Pillow==8.0.0
tensorflow==2.3.0